refactor(domain): BonusNumber의 getter 제거함
BonusNumber가 'getNumber()'를 노출하는 대신,
'validateDuplicate(WinningNumbers)' 메서드를 통해
스스로 당첨 번호와의 중복을 검증하도록 책임을 이동함.
